Reworked curated packages database and services
Moved IsLatest + IsLatestStable from CuratedPackage (which is essentially curated version of PackageRegistration) to it's own entity - CuratedPackageVersion (which is essentially curated version of Package). This mirrors the DB structure for the non-curated Package, which makes it a lot easier to query latest and latest stable curated packages when building/updating the Lucene index.

Fixes a couple of bugs:
* Curated packages could fall out of the Lucene index when any package in a registration was updated (e.g. download count) - the normal packages were always updated, but curated packages were not.
* The first time a curated package is created, the latest and latest stable flags were not being updated correctly. Subsequent curated package versions would be ok.